♻️ refactor(aiAgent): extract approvalResume pipeline + finalize execAgent skeleton
Part of LOBE-13706 (PR4 of 4, final). Completes the AiAgentService split:

- pipeline/approvalResume.ts: 2.6/2.7 approval claim + validation, the
  intervention-continuation reuse fast path, and the 16b/16c resume
  initialContext overrides
- pipeline/resolveRunAgentConfig.ts: stages 1-2.5 (config fetch, member
  overrides, model resolution, builtin runtime merge, scope injection)
- pipeline/turnSetup.ts: stage 3 topic creation/reuse + shared turn setup
  (device policy, hetero detection, attachment ingestion, user/assistant rows)
- pipeline/operationPrep.ts: stages 9.4-18 (device system info, agent
  management context, persona, history/message assembly, base initial context,
  workspace init, skill set, expertise) + the shared history loader
- pipeline/startOperation.ts: stage 19 (createOperation, reconnect marker,
  gateway token, startup-error handling)
- subAgentRuns.ts: execAgentThreadRun + execAgentMember as collaborators
- finalizeHeteroDispatchError internalized into pipeline/heteroDispatch.ts

execAgentWithReservation is now an explicit pipeline skeleton. index.ts
shrinks 4124 -> 1480 lines; largest file in the module is 1089 lines.
aiAgentService exports unchanged; the 24 existing __tests__ files
(287 tests) pass unmodified. Types + lint clean.
